refactor: move deck_text_cache from utils to repositories
The SQLite-backed deck-text cache is data-access-with-caching, which is
the Repositories layer's responsibility (cf. RadarRepository /
FormatCardPoolRepository, also SQLite). It was in utils only by inertia.
Move to repositories/deck_text_cache.py, export DeckTextCache /
get_deck_cache from repositories/__init__.py, and update the
navigators/services/test consumers.
